start mission6
